Add controller categoria
Se agrega controller de Categoria

+++ src/main/java/com/ljhernandez/shopAll/data/dto/ProductoDto.java
@@ -4,13 +4,15 @@ package com.ljhernandez.shopAll.data.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ProductoDto {
     private Integer idProdcuto;
     private String nombreProducto;
     private String descripcionProducto;
-    private double precioUnitario;
+    private BigDecimal precioUnitario;
     private int idCategoria;
     private Integer stockDisponible;
